+++ libraries/CHT8305/CHANGELOG.md
@@ -6,12 +6,24 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.0] - 2023-12-05
+- refactor API, constructor, begin()
+- update readme.md
+- rename setHumOffset => **setHumidityOffset()**
+- rename getHumOffset => **getHumidityOffset()**
+- rename setTempOffset => **setTemperatureOffset()**
+- rename getTempOffset => **getTemperatureOffset()**
+- update examples
+- update keywords.txt
+- minor edits
+
+----
+
 ## [0.1.7] - 2023-09-21
 - add Wire1 support for ESP32
 - fix humidity offset underflow + overflow.
 - update readme.md
 - minor performance optimization in humidity math
-
 
 ## [0.1.6] - 2023-02-01
 - update GitHub actions